% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_heatmap.R
\name{beta_heatmap}
\alias{beta_heatmap}
\title{β's Heatmap}
\usage{
beta_heatmap(
  object,
  scale = TRUE,
  scale_by = "model",
  low = "steelblue",
  high = "orange"
)
}
\arguments{
\item{object}{A LassoLoop object.}

\item{scale}{Logical indicating if the resultant heatmap will be scaled or not. Default is TRUE.}

\item{scale_by}{Character that indicates for which variable the resultant heatmap will be scaled. Options are "model" and "variable".}

\item{low}{Colour for the low end of the gradient. Lower beta values will be represented with this value.}

\item{high}{Colour for the high end of the gradient. Highest beta values will be represented with this value.}
}
\value{
A ggplot2 object.
}
\description{
Classical heatmap made with the beta values of each model in the "LassoLoop" object. This plot shows in a fast and intuitive way the direction of coefficients for each variable. Each heatmap row denotes one LassoLoop model and each column denotes one selected variable.
}
\author{
Pol Castellano-Escuder
}
