% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blasso.R
\name{blasso}
\alias{blasso}
\title{Bootstrap Validation for Quantitative Lasso Regression}
\usage{
blasso(
  x,
  y,
  loops = 2,
  bootstrap = TRUE,
  alpha = 1,
  nfolds = 10,
  offset = NULL,
  family = "gaussian",
  ntest = NULL,
  seed = 987654321,
  ncores = 2
)
}
\arguments{
\item{x}{x matrix as in glmnet.}

\item{y}{Response variable. Should be numeric a vector.}

\item{loops}{Number of loops (a \code{glmnet::cv.glmnet} model will be performed in each loop).}

\item{bootstrap}{Logical indicating if bootstrap will be performed or not.}

\item{alpha}{The elasticnet mixing parameter, with 0 ≤ alpha ≤ 1. alpha = 1 is the lasso penalty, and alpha = 0 the ridge penalty.}

\item{nfolds}{Number of folds - default is 10. Although nfolds can be as large as the sample size (leave-one-out CV), it is not recommended for large datasets. Smallest value allowable is nfolds=3.}

\item{offset}{A vector of length \code{nobs} that is included in the linear predictor. See \code{?glmnet::glmnet()}}

\item{family}{Response type. Quantitative for family = "gaussian" or family = "poisson" (non-negative counts).}

\item{ntest}{Numeric indicating the percentage of observations that will be used as test set. Default is NULL (no test set).}

\item{seed}{\code{set.seed()} that will be used.}

\item{ncores}{Number of cores. Each loop will run in one core using the \code{foreach} package.}
}
\value{
A LassoLoop object with the results.
}
\description{
This function performs n \code{glmnet::cv.glmnet(family = c("gaussian", "poisson"))} models using bootstrap validation and splitting the input data in train and test at each loop.
}
\references{
Jerome Friedman, Trevor Hastie, Robert Tibshirani (2010). Regularization Paths for Generalized Linear Models via Coordinate Descent. Journal of Statistical Software, 33(1), 1-22. URL http://www.jstatsoft.org/v33/i01/.
}
\author{
Pol Castellano-Escuder
}
