% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_heatmap.R
\name{coeff_heatmap}
\alias{coeff_heatmap}
\title{Title}
\usage{
coeff_heatmap(
  object,
  scale = TRUE,
  scale_by = "model",
  low = "steelblue",
  high = "orange"
)
}
\arguments{
\item{object}{A LassoLoop object.}

\item{scale}{f}

\item{scale_by}{f}

\item{low}{f}

\item{high}{f}
}
\value{

}
\description{

}
\author{
Pol Castellano-Escuder
}
