% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_blasso.R
\name{cox_blasso}
\alias{cox_blasso}
\title{Bootstrap Validation for Cox \code{glmnet::cv.glmnet} Function}
\usage{
cox_blasso(
  x,
  y,
  loops = 2,
  alpha = 1,
  nfolds = 10,
  seed = 987654321,
  ncores = 2
)
}
\arguments{
\item{x}{x matrix as in glmnet.}

\item{y}{Should be a two-column matrix with columns named 'time' and 'status'. The latter is a binary variable, with '1' indicating death, and '0' indicating right censored. The function Surv() in package survival produces such a matrix.}

\item{loops}{Number of resamplings (a \code{glmnet::cv.glmnet} model will be created at each resampling).}

\item{alpha}{The elasticnet mixing parameter, with 0 ≤ alpha ≤ 1. alpha = 1 is the lasso penalty, and alpha = 0 the ridge penalty.}

\item{nfolds}{number of folds - default is 10. Although nfolds can be as large as the sample size (leave-one-out CV), it is not recommended for large datasets. Smallest value allowable is nfolds=3.}

\item{seed}{\code{set.seed()} that will be used.}

\item{ncores}{Number of cores. Each loop will run in one core using the \code{foreach} package.}
}
\value{
A list with the results.
}
\description{

}
\references{
Jerome Friedman, Trevor Hastie, Robert Tibshirani (2010). Regularization Paths for Generalized Linear Models via Coordinate Descent. Journal of Statistical Software, 33(1), 1-22. URL http://www.jstatsoft.org/v33/i01/.
}
\author{
Pol Castellano-Escuder
}
