% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_binom_blasso.R
\name{random_binom_blasso}
\alias{random_binom_blasso}
\title{Bootstrap Validation for Binomial Random Lasso Regression}
\usage{
random_binom_blasso(
  x,
  y,
  loops = 2,
  random_vars = floor(sqrt(ncol(x))),
  bootstrap = TRUE,
  smote = FALSE,
  perc_over = 2,
  perc_under = 2,
  alpha = 1,
  nfolds = 10,
  seed = 987654321,
  ncores = 2
)
}
\arguments{
\item{x}{x matrix as in glmnet.}

\item{y}{Should be either a numeric factor with two levels.}

\item{loops}{Number of loops (a \code{glmnet::cv.glmnet} model will be performed in each loop).}

\item{random_vars}{Number of variables randomly sampled as candidates at each loop. Default is \code{sqrt(p)} (where p is number of variables in x).}

\item{bootstrap}{Logical indicating if bootstrap will be performed or not.}

\item{smote}{Logical. If it's set to TRUE, the Synthetic Minority Over-sampling Technique will be used to reduce random oversampling. See \code{performanceEstimation::smote} function.}

\item{perc_over}{If smote parameter is TRUE. A number that drives the decision of how many extra cases from the minority class are generated (known as over-sampling).}

\item{perc_under}{If smote parameter is TRUE. A number that drives the decision of how many extra cases from the majority classes are selected for each case generated from the minority class (known as under-sampling).}

\item{alpha}{The elasticnet mixing parameter, with 0 ≤ alpha ≤ 1. alpha = 1 is the lasso penalty, and alpha = 0 the ridge penalty.}

\item{nfolds}{number of folds - default is 10. Although nfolds can be as large as the sample size (leave-one-out CV), it is not recommended for large datasets. Smallest value allowable is nfolds=3.}

\item{seed}{\code{set.seed()} that will be used.}

\item{ncores}{Number of cores. Each loop will run in one core using the \code{foreach} package.}
}
\value{
A LassoLoop object with the results.
}
\description{
This function performs n \code{glmnet::cv.glmnet(family = "binomial")} models using a random subset of variables at each loop. Bootstrap validation will be used at each loop.
}
\references{
Jerome Friedman, Trevor Hastie, Robert Tibshirani (2010). Regularization Paths for Generalized Linear Models via Coordinate Descent. Journal of Statistical Software, 33(1), 1-22. URL http://www.jstatsoft.org/v33/i01/.
}
\author{
Pol Castellano-Escuder
}
