% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blasso.R
\name{blasso}
\alias{blasso}
\title{Title}
\usage{
blasso(
  x,
  y,
  loops = 2,
  nfolds = 10,
  family = "gaussian",
  lambda = NULL,
  seed = 987654321,
  ncores = 2
)
}
\arguments{
\item{ncores}{}
}
\value{

}
\description{

}
\references{

}
\author{

}
